+++ lib/container.dart
@@ -4,11 +4,10 @@ abstract class Container<T> {
   List<T> _children = new List<T>();
   Container() {}
 
-  void add(T node);
-
+  void addChild(T node);
   void removeChild(T node);
-
-  void insert(int index, T node);
+  void insertChild(int index, T node);
+  void clearChildren();
 
   List<T> get children;
 }